role merge security fix (be)

+++ src/main/java/com/example/auto_ria/controllers/user/UserController.java
@@ -1,20 +1,33 @@
 package com.example.auto_ria.controllers.user;
 
-import com.example.auto_ria.dto.updateDTO.UserUpdateDTO;
-import com.example.auto_ria.exceptions.CustomException;
-import com.example.auto_ria.models.responses.user.UserResponse;
-import com.example.auto_ria.models.user.AdministratorSQL;
-import com.example.auto_ria.models.user.ManagerSQL;
-import com.example.auto_ria.models.user.UserSQL;
-import com.example.auto_ria.services.CommonService;
-import com.example.auto_ria.services.user.UsersServiceMySQLImpl;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
+
+import com.example.auto_ria.dto.updateDTO.UserUpdateDTO;
+import com.example.auto_ria.enums.EMail;
+import com.example.auto_ria.enums.ERole;
+import com.example.auto_ria.exceptions.CustomException;
+import com.example.auto_ria.models.responses.user.UserResponse;
+import com.example.auto_ria.models.user.UserSQL;
+import com.example.auto_ria.services.CommonService;
+import com.example.auto_ria.services.user.UsersServiceMySQLImpl;
+
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.AllArgsConstructor;
 
 @RestController
 @AllArgsConstructor
@@ -24,10 +37,10 @@ public class UserController {
     private UsersServiceMySQLImpl usersServiceMySQL;
     private CommonService commonService;
 
+    @PreAuthorize("hasRole('ADMIN', 'MANAGER')")
     @GetMapping("/page/{page}")
     public ResponseEntity<Page<UserResponse>> getAll(
-            @PathVariable("page") int page
-    ) {
+            @PathVariable("page") int page) {
         try {
             return usersServiceMySQL.getAll(page);
         } catch (CustomException e) {
@@ -35,50 +48,113 @@ public class UserController {
         }
     }
 
+    @PreAuthorize("hasRole('ADMIN', 'MANAGER', 'USER')")
     @GetMapping("/{id}")
     public ResponseEntity<UserResponse> getById(@PathVariable("id") int id) {
         try {
             return usersServiceMySQL.getByIdAsResponse(id);
+
         } catch (CustomException e) {
             throw new CustomException(e.getMessage(), e.getStatus());
         }
     }
 
+    @PreAuthorize("hasRole('ADMIN', 'MANAGER', 'USER')")
     @PatchMapping("/{id}")
-    public ResponseEntity<UserSQL> patchUser(@PathVariable int id,
-                                                 @ModelAttribute UserUpdateDTO partialUser,
-                                                 HttpServletRequest request) {
+    public ResponseEntity<UserResponse> patchUser(@PathVariable int id,
+            @ModelAttribute UserUpdateDTO partialUser,
+            HttpServletRequest request) {
         try {
             UserSQL user = commonService.extractUserFromHeader(request);
             UserSQL userById = usersServiceMySQL.getById(id);
-            if (user != null && user.getId() != userById.getId()) {
-                throw new CustomException("Failed. Check credentials", HttpStatus.FORBIDDEN);
+
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+            if (authentication.getPrincipal() instanceof UserDetails) {
+                UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+                if (userDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ADMIN"))) {
+                    return usersServiceMySQL.update(id, partialUser, user);
+
+                } else if (userDetails.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals("MANAGER"))
+                        && userDetails.getUsername() == userById.getEmail()) {
+
+                    return usersServiceMySQL.update(id, partialUser, user);
+
+                } else if (userDetails.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals("USER"))
+                        && userDetails.getUsername() == userById.getEmail()) {
+
+                    return usersServiceMySQL.update(id, partialUser, user);
+
+                } else {
+                    throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
+                }
+            } else {
+                throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
             }
-            return usersServiceMySQL.update(id, partialUser, user);
+
         } catch (CustomException e) {
             throw new CustomException(e.getMessage(), e.getStatus());
         }
     }
 
+    @PreAuthorize("hasRole('ADMIN', 'MANAGER', 'USER')")
     @PatchMapping("/change-avatar/{id}")
     public ResponseEntity<String> patchAvatar(@PathVariable int id,
-                                              @RequestParam("avatar") MultipartFile avatar,
-                                              HttpServletRequest request) {
+            @RequestParam("avatar") MultipartFile avatar,
+            HttpServletRequest request) {
         try {
-            AdministratorSQL administrator = commonService.extractAdminFromHeader(request); //* todo has role adm or user check */
-            UserSQL user = commonService.extractUserFromHeader(request);
-            assert user != null;
-            if (user.getId() != id || administrator == null) {
-                throw new CustomException("Illegal_access_exception. No-permission: check credentials", HttpStatus.FORBIDDEN);
+            UserSQL userById = usersServiceMySQL.getById(id);
+
+            assert userById != null;
+
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+            if (authentication.getPrincipal() instanceof UserDetails) {
+                UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+                if (userDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ADMIN"))) {
+                    commonService.removeAvatar(userById.getAvatar());
+
+                    String fileName = avatar.getOriginalFilename();
+                    usersServiceMySQL.transferAvatar(avatar, fileName);
+                    usersServiceMySQL.updateAvatar(id, fileName);
+
+                    return ResponseEntity.ok("Success. Avatar_updated");
+
+                } else if (userDetails.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals("MANAGER"))
+                        && userDetails.getUsername() == userById.getEmail()) {
+
+                    commonService.removeAvatar(userById.getAvatar());
+
+                    String fileName = avatar.getOriginalFilename();
+                    usersServiceMySQL.transferAvatar(avatar, fileName);
+                    usersServiceMySQL.updateAvatar(id, fileName);
+
+                    return ResponseEntity.ok("Success. Avatar_updated");
+
+                } else if (userDetails.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals("USER"))
+                        && userDetails.getUsername() == userById.getEmail()) {
+
+                    commonService.removeAvatar(userById.getAvatar());
+
+                    String fileName = avatar.getOriginalFilename();
+                    usersServiceMySQL.transferAvatar(avatar, fileName);
+                    usersServiceMySQL.updateAvatar(id, fileName);
+
+                    return ResponseEntity.ok("Success. Avatar_updated");
+
+                } else {
+                    throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
+                }
+            } else {
+                throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
             }
 
-            commonService.removeAvatar(user.getAvatar());
-
-            String fileName = avatar.getOriginalFilename();
-            usersServiceMySQL.transferAvatar(avatar, fileName);
-            usersServiceMySQL.updateAvatar(id, fileName);
-
-            return ResponseEntity.ok("Success. Avatar_updated");
         } catch (CustomException e) {
             throw new CustomException(e.getMessage(), e.getStatus());
         }
@@ -87,19 +163,50 @@ public class UserController {
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteById(@PathVariable String id, HttpServletRequest request) {
         try {
-            UserSQL user = commonService.extractUserFromHeader(request);
-            ManagerSQL manager = commonService.extractManagerFromHeader(request);
-            AdministratorSQL administrator = commonService.extractAdminFromHeader(request); //todo separate or has role
+            UserSQL user = usersServiceMySQL.getById(Integer.valueOf(id));
 
-            if (administrator == null && manager == null) {
-                if (user == null || !Integer.valueOf(id).equals(user.getId())) {
-                    throw new CustomException("Illegal_access_exception. No-permission", HttpStatus.FORBIDDEN);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+            if (authentication.getPrincipal() instanceof UserDetails) {
+                UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+                if (userDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ADMIN"))) {
+
+                    commonService.removeAvatar(user.getAvatar());
+
+                    return usersServiceMySQL.deleteById(id, user, EMail.YOUR_ACCOUNT_BANNED);
+
+                } else if (userDetails.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals("MANAGER"))) {
+                    if (userDetails.getUsername() == user.getEmail()) {
+                        commonService.removeAvatar(user.getAvatar());
+
+                        return usersServiceMySQL.deleteById(id, user, EMail.PLATFORM_LEAVE);
+                    } else if (userDetails.getUsername() != user.getEmail()
+                            && user.getAuthorities().contains(ERole.USER)) {
+
+                        commonService.removeAvatar(user.getAvatar());
+                        return usersServiceMySQL.deleteById(id, user, EMail.YOUR_ACCOUNT_BANNED);
+
+                    } else {
+                        throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
+                    }
+
+                } else if (userDetails.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals("USER"))
+                        && userDetails.getUsername() == user.getEmail()) {
+
+                    commonService.removeAvatar(user.getAvatar());
+
+                    return usersServiceMySQL.deleteById(id, user, EMail.PLATFORM_LEAVE);
+
+                } else {
+                    throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
                 }
+            } else {
+                throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
             }
 
-            commonService.removeAvatar(user.getAvatar());
-
-            return usersServiceMySQL.deleteById(id, user, administrator, manager);
         } catch (CustomException e) {
             throw new CustomException(e.getMessage(), e.getStatus());
         }
